# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: Voucher Partner API
  version: 1.0.1
servers:
  - url: 'http://localhost:8087/api'
    description: Development server
  - url: 'https://localhost:8087/api'
    description: Development server with TLS Profile
paths:
...
  /batch/list/{period}:
    get:
      tags:
        - batch
      summary: Retrieves list of batches
      operationId: batchList
      description: Returns a list of batches
      parameters:
        - name: period
          in: path
          required: true
          description: The period in number of days in the past to which the query should be limited
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
  /batch/status/{batchId}:
    put:
      tags:
        - batch
      summary: Close a batch or mark it as deleted.
      operationId: batchStatus
      description: Close a batch or mark it as deleted.
      parameters:
        - name: batchId
          in: path
          required: true
          description: batch Identifier
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchStatusRequest'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatusResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
  /batch/details/{batchId}:
    get:
      tags:
        - batch
      summary: Returns the details of a specific batch
      operationId: batchDetails
      description: Returns the details of a specific batch
      parameters:
        - name: batchId
          in: path
          required: true
          description: Batch identifier
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDetailsResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    BatchListResponse:
      type: array
      items:
        $ref: '#/components/schemas/BatchListResponseObject'
    BatchListResponseObject:
      type: object
      properties:
        Id:
          type: integer
          description: Batch Identifier
        batchName:
          type: string
          description: The unique name of the batch
        batchComment:
          type: string
          description: A descriptive comment for the batch
        createDate:
          description: Batch creation date
          type: string
          format: date-time
        loadDate:
          description: The date when the loading was completed
          type: string
          format: date-time
        status:
          type: string
          description: The status of the batch. Can be one of:(O)pen, (A)ctivated, (D)eleted
        createUser:
          type: string
          description: The user who initially created the batch
        activateUser:
          type: string
          description: The user who activated the batch
    BatchDetailsResponse:
      type: object
      properties:
        Id:
          type: integer
          description: The unique id of the batch
        batchName:
          type: string
          description: The unique id of the batch
        batchComment:
          type: string
          description: A descriptive comment for the batch
        createDate:
          type: string
          format: date-time
          description: The date that the batch was created
        loadDate:
          type: string
          format: date-time
          description: The date when the loading was completed
        status:
          type: string
          description: The status of the batch. Can be one of
            (O)pen, (A)ctivated, (D)eleted
        createUser:
          type: string
          description: The user who initially created the batch
        activateUser:
          type: string
          description: The user who activated the batch
        deleteUser:
          type: string
          description: The user who deleted the batch, if this was the case
        deleteDate:
          type: string
          format: date-time
          description: The date when the batch was deleted, if this was the case
        products:
          type: array
          items:
            $ref: '#/components/schemas/BatchDetailsResponseObject'
    BatchDetailsResponseObject:
      type: object
      properties:
        productId:
          type: string
          description: product id
        productType:
          type: string
          description: quantities of vouchers linked to batch
        productDescription:
          type: string
          description: description for a specific prduct
        voucherDescription:
          type: string
          description: description for a voucher
        fileName:
          type: string
          description: name of the given file
        quantity:
          type: integer
          description: quantities of vouchers linked to batch
        startDate:
          type: string
          format: date-time
          description: start date
        endDate:
          type: string
          format: date-time
          description: end date
        voucherExpiryDate:
          type: string
          format: date-time
          description: voucher expiry date
...
    BatchStatusRequest:
      type: object
      properties:
        status:
          type: string
          description: Incoming status of a batch
        userName:
          type: string
          description: Vodacom username
    BatchStatusResponse:
      type: object
      properties:
        status:
          type: string
          description: New updated response
...
  securitySchemes:
    jwt:
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
security:
  - jwt: []